refactoring, fixed stack overflow error during json serialization with annotations
