+++ src/main/java/com/entity/Coordinates.java
@@ -1,4 +1,4 @@
-package com;
+package com.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -7,16 +7,12 @@ import javax.persistence.Id;
 
 import lombok.Data;
 
-@Entity
 @Data
-public class Login {
+@Entity
+public class Coordinates {
 	@Column(unique=true, nullable=false, updatable=false)
 	private @Id @GeneratedValue Long id;
-	private String uuid;
-	private String username;
-	private String password;
-	private String salt;
-	private String md5;
-	private String sha1;
-	private String sha256;
+	
+	private String latitude;
+	private String longitude;
 }
