+++ src/main/java/com/entity/Location.java
@@ -1,4 +1,4 @@
-package com;
+package com.entity;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
