+++ src/main/java/com/entity/Login.java
@@ -1,4 +1,4 @@
-package com;
+package com.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,11 +9,14 @@ import lombok.Data;
 
 @Entity
 @Data
-public class UserIdentity {
+public class Login {
 	@Column(unique=true, nullable=false, updatable=false)
 	private @Id @GeneratedValue Long id;
-	
-	private String name;
-	
-	private String value;
+	private String uuid;
+	private String username;
+	private String password;
+	private String salt;
+	private String md5;
+	private String sha1;
+	private String sha256;
 }
