+++ src/main/java/com/entity/Timezone.java
@@ -1,4 +1,4 @@
-package com;
+package com.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -7,14 +7,13 @@ import javax.persistence.Id;
 
 import lombok.Data;
 
-@Entity
 @Data
-public class DOB {
+@Entity
+public class Timezone {
 	@Column(unique=true, nullable=false, updatable=false)
 	private @Id @GeneratedValue Long id;
 	
-	private String date;
-	
-	private Integer age;
-
+	@Column(name="Timezone_Offset")
+	private String offset;
+	private String description;
 }
