+++ src/main/java/com/entity/UserIdentity.java
@@ -1,4 +1,4 @@
-package com;
+package com.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -7,13 +7,13 @@ import javax.persistence.Id;
 
 import lombok.Data;
 
-@Data
 @Entity
-public class Timezone {
+@Data
+public class UserIdentity {
 	@Column(unique=true, nullable=false, updatable=false)
 	private @Id @GeneratedValue Long id;
 	
-	@Column(name="Timezone_Offset")
-	private String offset;
-	private String description;
+	private String name;
+	
+	private String value;
 }
