+++ src/main/java/com/repository/NameRepository.java
@@ -1,0 +1,8 @@
+package com.repository;
+
+import com.Name;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NameRepository extends JpaRepository<Name, Long> {
+	
+}
