+++ src/main/java/com/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com;
+package com.repository;
 
+import com.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
