+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.username=sa
+spring.datasource.password=password
 
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
